Fix retrieving VHDX block size
We need the VHDX block size when retrieving the maximum internal
size of a VHDX for a specific file size.

The issue is that we rely on the virtdisk.dll API to give us this
value, while for fixed size images, the block size will always be
reported as 0.

This can easily be avoided by reintroducing the method which actually
parses the image file in order to retrieve the VHDX block size.

Closes-Bug: #1569451
Change-Id: Id035dbc070566817a60eed8270e39c52d3b54821