[wip - needs UT] Fix cluster group migration polling
We incorrectly assumed that the group state will immediately become
'Pending' when a migration is requested, judging by the expected
'ErrorPending' return code of the move function.

This is not the case, as the group can be reported as 'Online' while
being on the source node, before a live migration is actually
performed successfully. This is incorrectly treated as an exception
at the moment.

This change refactors the way in which we check pending migrations.
We're now waiting for cluster group change events, having 60 seconds
as a default migration timeout (which should be enough in most cases).

If a migration is attempted but fails, in most cases we'll get the
'Pending' state transition, quickly erroring out when the group state
changes without reaching the desired node/state.

In case of migration timeouts, we're now attempting to cancel pending
migrations.

In the future, we're planning to implement a custom cluster resource
dll, having a finer control over this.

Closes-Bug: #1628938

Change-Id: I6c268355213718d4d66e0e77a89c558cde3213b6