[wip] Add processutils module
This change adds a processutils module, which includes several
process/job helpers.

Associating a job with a process will allow ensuring that when a
parent process is killed, process associated with jobs (e.g. child
processes) will be killed as well.

This is especially useful for Cinder, which will use it for the
multiple backend per service scenario.

Change-Id: Icc2bbe9191e6db685c0fd294abc1d0eb24bc420c