[Trivial] Allow flushing named pipe log handles
The named pipe handler provided by os-win is used by Nova when
consuming instance serial ports. For convenience, it allows logging
the named pipe output to a file (useful for storing the serial
console output).

The issue is that the log file never gets flushed, thus Nova returns
incomplete results. This change adds a method that may be used for
flushing the log file.

Closes-Bug: #1741494