WIP: Add mutex class
This will expose Windows mutexes. We usually rely on file locks
when synchronizing multiple processes, mostly for portability
reasons.

Some times, file locks may not be an option, in which case a
named mutex can be useful. We intend to use this in Glance when
ensuring that only one scrubber can run at a time.

Change-Id: I4a2a51e2164c6f5816f2af2641d41d1db7870600